Fix junit.xml generation
